Rework how workloads are define
In this commit I create a new Type ITContract that allows to create new
workloads type more easily. To make a type a workloads you now just need
to embed the ITContract struct into your type and that's it.

ITContract also implemets the Workloader interface so all the code of
the explorer can deal with what is needs in an easy way.