Mark nodes as deleted instead of actually deleting
From now on, set a tombstone marker on nodes which are deleted. A new
filter is added when listing nodes, to also show deleted nodes (which is
not done by default). Updates to nodes marked as deleted are also
prohibited (the node will not be found). However, to maintain old
behavior, the NodeCreate call _is_ allowed (and will now clear the
tombtone if present). Looking up  a specific node will always succeed.

Fix #164